add rain data and fix web-app errors

+++ hardware/include/payload.hpp
@@ -9,7 +9,7 @@ class Payload {
     byte soil;
     int temperature;
     int pressure;
-    byte light;
+    int light;
     bool rain;
     dht11 dht;
